feat: extend Hash class to allow deep slice and use it in the rst formatter module
